4.3.4-dev-16

+++ src/phase2.h
@@ -1,6 +1,6 @@
 #ifndef PHASE2_H
 #define PHASE2_H
 
-double solveNLEforMasses(nle_config_t *nle_config, nle_state_t *nle_state);
+long double solveNLEforMasses(nle_config_t *nle_config, nle_state_t *nle_state);
 
 #endif // PHASE2_H
